handling of selection messages in tables
git-svn-id: https://svn.52north.org/svn/ilwis/drawers@1905 b64c5c0c-1e74-440b-9268-4ad09cf26390

+++ Sources/Client/Editors/LinearFilterView.h
@@ -59,7 +59,7 @@ public:
 	LinearFilterView();           // protected constructor used by dynamic creation
 	virtual ~LinearFilterView();
 	virtual void InitColPix(CDC* cdc);
-  virtual int iCols() const;
+  virtual long iCols() const;
   virtual long iRows() const;
   virtual String sColButton(int iCol) const;
   virtual String sRowButton(long iRow) const;
