+++ Sources/Client/Editors/MatrixView.h
@@ -60,7 +60,7 @@ public:
 	virtual ~MatrixView();
 
 	virtual void InitColPix(CDC* cdc);
-  virtual int iCols() const;
+  virtual long iCols() const;
   virtual long iRows() const;
   virtual String sColButton(int iCol) const;
   virtual String sRowButton(long iRow) const;
